.......PS. [ZBX-15238] casting time_t to int for use with %d specifier

+++ src/zabbix_server/discoverer/discoverer.c
@@ -719,7 +719,7 @@ static int	process_discovery(void)
 				" and r.nextcheck<=%d"
 				" and " ZBX_SQL_MOD(r.druleid,%d) "=%d",
 			DRULE_STATUS_MONITORED,
-			time(NULL),
+			(int)time(NULL),
 			CONFIG_DISCOVERER_FORKS,
 			process_num - 1);
